MRD-107: Make the start scripts work for M1 macs
This adds options to control when to...

- Run `docker-compose pull` (this isn't needed all the time)
- Build `hmpps-auth` (only on M1 macs)

The `hmpps-auth` build step is needed at least once if you're running an M1 mac as the currently distributed container isn't multi arch and has running problems on M1 silicon.

It'll also rebuild `hmpps-auth` if you run `docker-compose pull` as you'll have re-downloaded the "broken"  image.