MRD-2529 Fix OSP/C & OSP/DC VERY_HIGH display
The check on the level value was incorrect for OSP/C and OSP/DC values in the
case of a value of VERY_HIGH: the value itself is 'VERY_HIGH', with an
underscore, while the value checked against was 'VERY HIGH', with a space in
between. This was resulting in the label marker for VERY_HIGH cases to be
displayed in the middle of the rating bar, rather than on the right-most
section, due to the label not having the right class assigned to it. This is
corrected here.