MRD-2611 Fix autocomplete selection issue
The autocomplete field wasn't behaving as we want it to: if a valid option was
set at any point, the component would remember it even if an invalid value was
provided afterwards (e.g. select a valid option, then type in an alternative
search which doesn't produce any results). This resulted in the rememberedç
value being set if the relevant form was submitted, even if it wasn't being
displayed to the user at the time.

This has been fixed by checking every time the field loses focus: if what is
currently typed into the field doesn't correspond to any of the available
options, the empty value option will be selected in the underlying select
element.