MRD-2611 Hide autocomplete placeholder on click
The autocomplete placeholder doesn't work well in the autocomplete component if
combined with the 'autoselect' option. With the latter configured, as soon as
you click on the input field it will display in a faint font which option will
be set if the 'Enter' key is pressed (if you set a minimum query length I
expect this issue doesn't occurs, but we aren't setting a minimum length).
Since the placeholder doesn't vanish until you start typing, the placeholder
text and that coming from the autoselect functionality overlap in the same
field.

This is addressed here via CSS: whenever the placeholder is focused on, it is
made invisible.