MRD-2639 Update recommendation.spec.ts
The test for the new 'Edit current establishment' page has been corrected to
only mock what it strictly needs to avoid misinterpretations of what the page
should do. Other tests in this file can also do with this stripping out too,
but will be done separately.

Additionally, a copy of the STATUSES enum defined in
recommendationStatusCheck.ts has been made in a new recommendationStatus.ts
file and recommendation.spec.ts has been updated to use it. Attempts were made
to import the existing enum, but compilation issues emerged whose cause could
not be pinpointed exactly (probably related to what else is being brought in as
part of the import). Since the enum isn't used in recommendationStatusCheck.ts,
it makes sense anyway to move it out. This move hasn't been done, and a copy
made instead, because a lot of other files would need to have been updated,
which is out of the scope we want to keep this set of changes to. A separate
ticket (MRD-2688) has been opened to make this move.