MRD-2639 Add current establishment to PPCS journey (#1359)
When a PPCS user books an offender into PPUD, they fill in the offender's
current establishment. At the moment, this happens manually, but it is going to
be incorporated into the CaR journey, so we need to make these changes here so
the establishment can be stored during said journey and passed through to PPUD
at booking time.

Our PPUD Automation API application removes 'Not Specified' options from the
reference lists it gets from PPUD. This includes the one with the list of
establishments, so the code defaults to blank in cases where 'Not Specified'
might seem appropriate, and the user is forced to pick an option. No cases were
described by PPCS where they would deliberately leave it blank during the booking
process, so this behaviour should be OK.