MRD-2700 Implement establishment mapping logic (#1364)
After discussions with PPCS, we are going with the following approach for the
mapping logic:

    Check NOMIS for the offender's prison.
    a. If the record has a prison set, take that and do the mapping to the
    relevant PPUD value.
    b. If it's OUT , the offender is UAL and we set Not Applicable.
    c. If it's TRN , go to step 2.
    Check NOMIS for the offender's movements.
    a. If the latest one is a transfer, take the destination prison.
    b. If it's something other than a transfer, go back to step 1 to see if the
    prison value has changed to something other than TRN . If it has changed,
    go with that; otherwise leave it blank for the user to check manually.

An alternative discussed to leaving the field blank at step 2.b was to just
take the latest transfer movement and go with that, even if it isn’t the latest
movement.

An alternative discussed to step 2 was to take the latest transfer (even if it
isn't the latest movement) and go with the destination prison there.

The pitfalls of these two alternatives (i.e. that another movement exists which
should be picked and which reflects a different establishment) are very
unlikely to happen, and there is also a chance that NOMIS data changes after it
is checked by CaR but before the PPCS user finishes processing the
recommendation anyway, but ultimately we decided it was better to leave the
field blank for the user to check manually. We expect this to happen very
rarely (if ever), so the disruption to the ideal flow of not having to access
NOMIS at all will very rarely (if ever) be experienced.

Regarding the PPCS user being made aware of the offender being in the middle of
a transfer: we will not be displaying this information on screen, as the only
case where it is ever relevant to them is for figuring out the establishment,
which we are automating here.