MRD-2700 Add missing API mock
The check-booking-details test is here updated to include an agencyId as part
of the offender's details, which in turn triggers the retrieval of the
establishment mappings in order to determine the relevant PPUD establishment.
Although not strictly necessary for the page to load, the most common case will
be for there to be an agencyId and for it to be mapped, so we exercise that
code here.