MRD-2684 Support indeterminate sentence booking
The createOrUpdateSentence function is updated here to support updating
indeterminate sentences on PPUD. Indeterminate sentence creation does not need
to be supported for now.

These changes also update the PpudUpdateSentenceRequest type to be in line with
how it's changed in the API service, as many fields are only mandatory in one
of the two types of sentences.