MRD-2780 Add .nvmrc file to project (#1420)
The npm security scan requires the node version to be passed to it through a
file. It defaults to checking .nvmrc. We don't have any file with the version
at the moment, so we add an .nvmrc one for this. This should also make it
easier to switch to the right node version when working on the project.