MRD-2883 Group task list integration tests
The task list integration tests were timing out in CircleCI. This may be due to
the number of 'it' calls made, as the function was being called per check made
rather than per set-up. They have been grouped to have one 'it' call per set-up
to see if this will reduce the resource consumption and timings.