MRD-2883 Add vulnerabilities details page to the task list (#1436)
The splitting of the vulnerabilities page into separate selection and details
pages is here incorporated into the task list page (behind feature flag). The
task completion logic has also been adjusted accordingly, to ensure
vulnerability details are still required in order to proceed to countersignature
and Part A creation.

The vulnerabilities and victim liaison sections were included in the same njk
file to be included in the task list one, even though they are separate topics
and sections. The rest of the sections are already separated, each one in its
own njk, and there is no reason stated in the PR where they were put together
over why they should be in the same file, so I am splitting them here in order
to separate concerns.