Fix editReleaseDate.spec.ts
The editReleaseDate test was failing occasionally in an irreproducible manner.
It looks like the likely reason is that the test is assuming the local timezone
is Europe/London, while the production code doesn't and makes sure the date is
always printed for Europe/London. If the date randomly generated for the test
fell on different days in Europe/London vs the local timezone, the test would
fail. The test is here fixed to ensure it's checking for Europe/London.