Fix personSearch.spec.ts
The personSearch test was implicitly depending on other integration tests' mock
set-up. This only came to light when the integration tests were split to be run
in parallel in the pipeline. The test is corrected here to set up the mock it
needs. Ideally it should also reset said set-up at the end, as should the rest
of the integration tests, but this is left for a separate ticket (MRD-2944), as
it's its own can of worms and we're just looking to get the pipeline flowing
again for now.