Run integration tests in parallel (#1441)
We attempt to parallelise the integration test runs in CircleCI here. We cannot
use the Cypress CircleCI orb for this, as it only supports parallel runs with a
Cypress Cloud account. Instead, we do the splitting and parallel running
"manually":
 * circleci tests glob - this gets the full list of tests to run
 * circleci tests split - this splits the tests into as many chunks as
    parallel runs specified in the job's parallelism parameter. Note we are
    splitting by timing, which relies on having used the 'store_test_results'
    CircleCI command on previous runs.
 * paste -sd ',' - concatenates the tests, comma-separating them

The editReleaseDate test was failing occasionally in an irreproducible manner.
It looks like the likely reason is that the test is assuming the local timezone
is Europe/London, while the production code doesn't and makes sure the date is
always printed for Europe/London. If the date randomly generated for the test
fell on different days in Europe/London vs the local timezone, the test would
fail. The test is here fixed to ensure it's checking for Europe/London.

The personSearch test was implicitly depending on other integration tests' mock
set-up. This only came to light when the integration tests were split to be run
in parallel in the pipeline. The test is corrected here to set up the mock it
needs. Ideally it should also reset said set-up at the end, as should the rest
of the integration tests, but this is left for a separate ticket (MRD-2944), as
it's its own can of worms and we're just looking to get the pipeline flowing
again for now.