MRD-2778 Make PPUD indeterminate sentence fields optional (#1445)
Several fields in PpudSentenceData are made nullable here. Although we expect
the user to have provided values for all of the fields in order to book the
recall into PPUD, initially this object is filled in with data from PPUD,
where only the dateOfSentence is guaranteed to be present. The other fields
may be null if the PPUD data is incomplete.