Fix recallType.spec.ts test
One of the test cases in recallType.spec.ts was flawed: it uses a
randomiseCriteria function to generate a random set of boolean values, of which
it wants at least one to be true. The criteria passed to the method to check was
instead checking that they weren't all true, so occasionally the object
generated was all false values, which caused the tests to fail. This has now
been fixed to check the values aren't all false.