MRD-3007 Add npm-script-allowlist to project (#1466)
This set of changes adds npm-script-allowlist to the project to control which
of our dependencies run post-install scripts. This helps guard against attacks
seen recently relying on such scripts. It introduces a 'setup' script to run
instead of 'npm install' or 'npm ci', which we also use in our pipeline.

Related to this are the changes around 'sudo npm install' in the pipeline:
The SRE team have highlighted that using 'sudo npm install -g npm@<version>' is
both unsafe (unnecessary elevated privileges) and unnecessary, as there are
official actions/commands that can be used to install Node and the desired npm
version. Here we also deliberately align the two instances we had of the sudo
calls, as it's unlikely we still need to pin the npm version to 9 in the case
where that is currently done (it is unclear why it had to be pinned in the
first place). The node version we have specified in our .nvmrc file (which the
actions/commands pick up by default) comes with npm 10, which is what we want.